requirements.txt
fyers-apiv3
pandas
pandas_ta
scikit-learn
python-dotenv